refactor account margin
