+++ account.js
@@ -1,7 +1,20 @@
 // account.js
-import { getAccountMargin } from "./orderExecution.js";
+// Provides utilities for fetching and caching account related data
+import { kc, initSession } from "./kite.js";
 
 let accountBalance = 0;
+
+// Fetch margin available across equity using the shared Kite instance
+export async function getAccountMargin() {
+  try {
+    await initSession();
+    const response = await kc.getMargins("equity");
+    return response;
+  } catch (err) {
+    console.error(`[ACCOUNT] Error fetching account margin`, err?.message || err);
+    return null;
+  }
+}
 
 export function getAccountBalance() {
   return accountBalance;
